+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.7.0] - 2021-10-04
 ### Changed
 - The `Source` enum has a new variant `SharedFile`, used for unsafe persistent
   memory mappings.
@@ -68,7 +70,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Stretch processing. `ttf-parser` was incorrectly parsing this property.
 
-[Unreleased]: https://github.com/RazrFalcon/fontdb/compare/v0.6.2...HEAD
+[Unreleased]: https://github.com/RazrFalcon/fontdb/compare/v0.7.0...HEAD
+[0.7.0]: https://github.com/RazrFalcon/fontdb/compare/v0.6.2...v0.7.0
 [0.6.2]: https://github.com/RazrFalcon/fontdb/compare/v0.6.1...v0.6.2
 [0.6.1]: https://github.com/RazrFalcon/fontdb/compare/v0.6.0...v0.6.1
 [0.6.0]: https://github.com/RazrFalcon/fontdb/compare/v0.5.4...v0.6.0